+++ actix/CHANGES.md
@@ -1,6 +1,11 @@
 # CHANGES
 
 ## Unreleased - 2021-xx-xx
+### Added
+* Add `fut::try_future::ActorTryFuture`. [#419]
+* Add `fut::try_future::ActorTryFutureExt` trait with `map_ok`, `map_err` and `and_then` combinator. [#419] 
+
+[#419]: https://github.com/actix/actix/pull/419
 
 
 ## 0.11.1 - 2021-03-23